chore(overview): breakdown component charts
