+++ components/dialog-sql.tsx
@@ -1,13 +1,10 @@
 import { CodeIcon } from '@radix-ui/react-icons'
 
-import { DialogContent } from '@/components/dialog-content'
+import { DialogContent, DialogContentProps } from '@/components/dialog-content'
 import { Button } from '@/components/ui/button'
 import { dedent } from '@/lib/utils'
 
-interface ShowSQLButtonProps {
-  button?: React.ReactNode
-  title?: string
-  description?: string
+interface ShowSQLButtonProps extends Omit<DialogContentProps, 'content'> {
   sql?: string
 }
 
@@ -16,6 +13,7 @@ export function DialogSQL({
   title = 'SQL Code',
   description = 'Raw SQL code of this table',
   sql,
+  ...props
 }: ShowSQLButtonProps) {
   if (!sql) {
     return null
@@ -39,7 +37,8 @@ export function DialogSQL({
       }
       title={title}
       description={description}
-      content={<pre className="text-sm">{dedent(sql)}</pre>}
+      content={<pre className="text-wrap text-sm">{dedent(sql)}</pre>}
+      {...props}
     />
   )
 }
